Added files via upload
I abhor turning in incomplete work, but I can't seem to get rid of a few "expected a declaration" errors. I've tried a variety of ways, but I think this is the best I can do. The book was quite helpful in rectifying my pointer array problems. Sorry nothing runs.